+++ bin/drivers/install-geckodriver-arm.sh
@@ -1,6 +1,5 @@
 #!/usr/bin/env bash
-# https://raspberrypi.stackexchange.com/questions/63258/selenium-firefox-oserror-errno-8-exec-format-error
-wget "https://github.com/mozilla/geckodriver/releases/download/v0.19.1/geckodriver-v0.19.1-arm7hf.tar.gz" -O /tmp/geckodriver.tar.gz
-sudo tar -xvzf /tmp/geckodriver*
-sudo chmod +x ./geckodriver
-sudo mv ./geckodriver /usr/local/bin/
+geckodriver_version="0.35.0"
+wget "https://github.com/mozilla/geckodriver/releases/download/v$geckodriver_version/geckodriver-v$geckodriver_version-linux-aarch64.tar.gz"
+tar -xvf "geckodriver-v$geckodriver_version-linux-aarch64.tar.gz"
+sudo mv geckodriver /usr/local/bin/